Refactor the file by separating them into their own files
